Moving to webpacker 6 / webpack 5
