+++ app/javascript/doorkeeper.js
@@ -1,0 +1,3 @@
+import "./src/language";
+
+import "./stylesheets/doorkeeper.scss";
