+++ app/javascript/intro.js
@@ -1,0 +1,9 @@
+import "bootstrap";
+
+import "../../vendor/assets/javascripts/jquery.validate.min.js";
+import "./src/intl_tel_number.js";
+import "./src/language.js";
+import "./src/intro.js";
+
+require("@rails/ujs").start();
+require("turbolinks").start();
