+++ app/javascript/members.js
@@ -1,0 +1,12 @@
+import "bootstrap";
+
+require("@rails/ujs").start();
+require("turbolinks").start();
+
+import "./src/intl_tel_number";
+import "./src/members/main";
+import "./src/members/activities/activities";
+import "./src/members/activities/activity";
+import "./src/members/activities/poster_modal";
+import "./src/members/payments";
+import "./src/language";
