--- app/javascript/packs/doorkeeper.js
@@ -1,5 +1,0 @@
-import "bootstrap";
-
-import "../src/language";
-
-import "../stylesheets/doorkeeper.scss";
