+++ app/javascript/public.js
@@ -1,5 +1,5 @@
 import "bootstrap";
-import "../src/public.js";
+import "./src/public.js";
 
 require("@rails/ujs").start();
 require("turbolinks").start();
