--- app/views/admin/members/payment_whatsapp.html.erb
@@ -1,10 +1,0 @@
-<%= I18n.t('admin.payment.whatsapp.title', user: current_user.credentials.first_name) %>
-
-<%= I18n.t('admin.payment.whatsapp.open_activities') %>:
-<% @participants.each do |p| %>
-- <%= p.activity.name %> - <%= number_to_currency(p.currency, unit: '€') %>
-<% end %>
-
-<%= I18n.t('admin.payment.whatsapp.total_of', amount: number_to_currency(@participants.map(&:currency).sum, unit: '€')) %>
-
-<%= I18n.t('admin.payment.whatsapp.pay_to', member: @member.name) %>
