+++ config/webpack/development.js
@@ -1,9 +1,9 @@
 process.env.NODE_ENV = process.env.NODE_ENV || 'development'
 
 //const SpeedMeasurePlugin = require('speed-measure-webpack-plugin');
-const environment = require('./environment')
+const webpackConfig = require('./base')
 
 //const smp = new SpeedMeasurePlugin();
-const config = environment.toWebpackConfig();
+const config = webpackConfig;
 
 module.exports = config;
