+++ package.json
@@ -10,10 +10,11 @@
   ],
   "dependencies": {
     "@rails/ujs": "~6.0.3-4",
-    "@rails/webpacker": "~5.4.3",
+    "@rails/webpacker": "~6.0.0-rc.6",
     "autoprefixer": "~10.4.0",
     "bootstrap": "~4.3.1",
     "clipboard": "~1.6.1",
+    "css-loader": "~6.5.1",
     "i18n-js": "~3.8.0",
     "intl-tel-input": "~16.0.5",
     "jquery": "~3.5.1",
@@ -21,18 +22,28 @@
     "js-cookie": "~2.2.1",
     "popper.js": "~1.16.1",
     "postcss": "~8.3.11",
-    "postcss-loader": "~4.2.0",
+    "postcss-loader": "~6.2.0",
     "postcss-preset-env": "~6.7.0",
     "postcss-import": "~14.0.2",
     "postcss-flexbugs-fixes": "~5.0.2",
+    "mini-css-extract-plugin": "~2.4.4",
+    "css-minimizer-webpack-plugin": "~3.1.3",
+    "sass-loader": "~12.3.0",
+    "style-loader": "~3.3.1",
+    "sass": "~1.43.4",
     "rails-erb-loader": "~5.5.2",
     "sweetalert2": "~9.14.2",
     "tailwindcss": "~2.2.19",
     "toastr": "~2.1.4",
     "turbolinks": "~5.2.0",
-    "webpack-dev-server": "~3.11.0"
+    "webpack-dev-server": "~4.4.0"
   },
   "devDependencies": {
     "speed-measure-webpack-plugin": "~1.3.3"
+  },
+  "babel": {
+    "presets": [
+      "./node_modules/@rails/webpacker/package/babel/preset.js"
+    ]
   }
 }
