+++ postcss.config.js
@@ -3,12 +3,13 @@ module.exports = {
     require("postcss-import"),
     require("autoprefixer"),
     require("tailwindcss"),
-    require("postcss-flexbugs-fixes"),
     require("postcss-preset-env")({
       autoprefixer: {
         flexbox: "no-2009",
       },
       stage: 3,
     }),
+
+    require("postcss-flexbugs-fixes"),
   ],
 };
